% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{gen_chisq}
\alias{gen_chisq}
\title{Chi-Squared Points}
\usage{
gen_chisq(n, df)
}
\arguments{
\item{n}{Number of points to generate.}

\item{df}{Degrees of freedom.}
}
\value{
N-vector containing the sampled points.
}
\description{
This function generates points targeting a Chi-Squared
distribution.
}
\examples{
gen_chisq(n = 20, df = 3)
}
