% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{gen_exp}
\alias{gen_exp}
\title{Exponential Points}
\usage{
gen_exp(n, rate = 1)
}
\arguments{
\item{n}{Number of points to generate.}

\item{rate}{Rate parameter.}
}
\value{
N-vector containing the sampled points.
}
\description{
This function generates points targeting an Exponential
distribution.
}
\examples{
gen_exp(n = 15, rate = 1)
}
