% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{gen_gh}
\alias{gen_gh}
\title{Tukey's g-and-h Points}
\usage{
gen_gh(n, a = 0, b = 1, g = 0, h = 0)
}
\arguments{
\item{n}{Number of points to generate.}

\item{a}{Location parameter.}

\item{b}{Scale parameter.}

\item{g}{Symmetry parameter. The distribution is symmetric when g=0.}

\item{h}{Tail shape parameter.}
}
\value{
N-vector containing the sampled points.
}
\description{
This function generates points targeting a g-and-h
distribution.
}
\examples{
gen_gh(n = 20, a = 0, b = 1, g = 0, h = 1)
}
