% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{gen_gk}
\alias{gen_gk}
\title{g-and-k Points}
\usage{
gen_gk(n, a = 0, b = 1, g = 0, k = 0)
}
\arguments{
\item{n}{Number of points to generate.}

\item{a}{Location parameter.}

\item{b}{Scale parameter.}

\item{g}{Symmetry parameter. g = 0 yields a symmetric distribution,
g > 0 gives positive skewness, g < 0 gives negative skewness.}

\item{k}{Tail shape parameter. k > 0 gives longer tails than the normal.
Values of k < 0 give shorter tails.}
}
\value{
N-vector containing the sampled points.
}
\description{
This function generates points targeting a g-and-k
distribution.
}
\examples{
gen_gk(n = 20, a = 0, b = 1, g = 2, k = 0)
}
