% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{gen_norm}
\alias{gen_norm}
\title{Normal Points}
\usage{
gen_norm(n, mu = 0, sig = 1)
}
\arguments{
\item{n}{Number of points to generate.}

\item{mu}{Mean.}

\item{sig}{Standard deviation.}
}
\value{
N-vector containing the sampled points.
}
\description{
This function generates points targeting a normal
distribution.
}
\examples{
gen_norm(n = 20, mu = 2, sig = 3)
}
