% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{gen_unif}
\alias{gen_unif}
\title{Uniform Points}
\usage{
gen_unif(n, l = 0, u = 1)
}
\arguments{
\item{n}{Number of points to generate.}

\item{l}{Lower bound.}

\item{u}{Upper bound.}
}
\value{
N-vector containing the sampled points.
}
\description{
This function generates points targeting a uniform
distribution.
}
\examples{
gen_unif(n = 10, l = 2, u = 5)
}
